+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - 0.1.0 (not yet released)
+  - Added `parallelMap()` and `makeParallelMap()`.
   - Added `makeAsync()`.
 
 - 0.0.0 (2012/02/19)